bots/extrator/requirements.txt
playwright==1.49.1
gspread==6.1.4
google-auth==2.37.0
python-dotenv==1.0.1
pydantic==2.10.3
